pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "podcast-stats"
version = "0.1.0"
description = "Podcast transcription system with speaker diarization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "pyannote.audio>=3.1",
    "openai>=1.0.0",
    "torch>=2.0",
    "pydub>=0.25",
    "click>=8.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.scripts]
rt-audio = "src.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
